pyproject.toml
[tool.poetry]
name = "cleantoots"
version = "0.3.1"
description = "Cleanup your toot history."
license = "GPL-3.0-or-later"
authors = ["Gabriel Augendre <gabriel@augendre.info>"]
readme = "README.md"
homepage = "https://github.com/Crocmagnon/cleantoots/tree/master"
keywords = ["mastodon", "toot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Communications :: Chat",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = ">=3.6"
"Mastodon.py" = "^1.5.0"
click = "^7.0"
pendulum = "^2.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[tool.poetry.scripts]
cleantoots = 'cleantoots.main:cli'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
